news/backend_server/requirement.txt
astroid==2.1.0
isort==4.3.4
jsonrpclib-pelix==0.4.0
lazy-object-proxy==1.3.1
mccabe==0.6.1
pylint==2.2.2
pymongo==3.7.2
six==1.12.0
wrapt==1.11.1
